Change namespace of plotting functions.
