+++ src/ttsim/plot/__init__.py
@@ -1,0 +1,3 @@
+from ttsim.plot import dag
+
+__all__ = ["dag"]
