+++ src/ttsim/plot/dag/__init__.py
@@ -1,0 +1,4 @@
+from ttsim.plot.dag.interface import interface
+from ttsim.plot.dag.tt import NodeSelector, tt
+
+__all__ = ["NodeSelector", "interface", "tt"]
