+++ src/ttsim/testing_utils.py
@@ -16,7 +16,7 @@ from ttsim.interface_dag_elements.data_converters import (
 )
 from ttsim.interface_dag_elements.fail_if import format_list_linewise
 from ttsim.interface_dag_elements.shared import to_datetime
-from ttsim.plot_dag import convert_all_nodes_to_callables
+from ttsim.plot.dag.tt import convert_all_nodes_to_callables
 from ttsim.tt.column_objects_param_function import PolicyInput
 
 # Set display options to show all columns without truncation
